requirements.txt
google-api-python-client
httplib2
python-gflags
